Add files via upload
This project is PURELY HTML.The project has a home page, about us page, order page, shipping information page and contact us page. Each category has four sample images and their description and prices displayed underneath it. This is still work-in-progress as I can and will in future add a login page and a page with order summary and a few more things. Also, this page does not have the option to select sizes so, there is a room for improvement.